+++ source/runner/_tests/fixtures/__DangerfileDefaultExportAsync.js
@@ -1,0 +1,7 @@
+export default async () =>
+  new Promise(res => {
+    setTimeout(() => {
+      warn("Asynchronous Warning")
+      res()
+    }, 10)
+  })